Improve code organization and transformer handling
Summary
* Move transformer logic to separate factory class
* Clean up server.ts to focus on HTTP endpoints
* Improve error handling and response formats

Details
* Create transformer-factory.ts with transform functions and factory class
* Remove transform logic from server.ts
* Add proper error responses for missing transformer/address
* Make server port configurable via environment variable
* Standardize JSON response formats
* Add factory methods for different transformer types
