+++ osc-listener/src/config.ts
@@ -9,6 +9,7 @@ export interface Config {
     localAddress: string;
     localPort: number;
     updateRate: number; // in Hz
+    serverPort: number;
     debug: DebugLevel;
 }
 
@@ -16,5 +17,6 @@ export const defaultConfig: Config = {
     localAddress: '0.0.0.0',
     localPort: 9005,
     updateRate: 1,
+    serverPort: 3001,
     debug: DebugLevel.Medium
 };
